@@ -10,7 +10,7 @@ import com.github.zavier.dto.ExpenseRecordUpdateCmd;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 import java.util.Objects;
 import java.util.Optional;
 
