@@ -7,7 +7,7 @@ import com.github.zavier.dto.ProjectAddCmd;
 import com.github.zavier.project.executor.converter.ProjectConverter;
 import org.springframework.stereotype.Component;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 
 @Component
 public class ProjectAddCmdExe {
