@@ -8,7 +8,7 @@ import com.github.zavier.dto.ProjectListQry;
 import com.github.zavier.dto.data.ProjectDTO;
 import org.springframework.stereotype.Component;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 import java.util.List;
 import java.util.stream.Collectors;
 
