@@ -11,7 +11,7 @@ import com.github.zavier.dto.data.statistics.PieStatisticsDTO;
 import com.github.zavier.utils.FreemarkerUtils;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.*;
