@@ -4,7 +4,7 @@ import com.alibaba.cola.dto.PageResponse;
 import com.github.zavier.domain.user.User;
 import com.github.zavier.dto.UserListQry;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 import java.util.Optional;
 
 public interface UserGateway {
